pyproject.toml
# Copyright AGNTCY Contributors (https://github.com/agntcy)
# SPDX-License-Identifier: Apache-2.0

[project]
name = "slima2a"
version = "0.4.0"
description = "A2A protocol over slimrpc"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "a2a-sdk[telemetry,sqlite]==1.0.0-alpha.0",
    "slim-bindings~=1.1",
]
classifiers = [
    "Development Status :: 3 - Alpha",

    # Indicate who your project is intended for
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",

    # Specify the Python versions you support here.
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Repository = "https://github.com/agntcy/slim"
Issues = "https://github.com/agntcy/slim/issues"
Changelog = "https://github.com/agntcy/slim/blob/main/data-plane/python/integrations/slima2a/CHANGELOG.md"

[dependency-groups]
linting = ["ruff>=0.12", "mypy>=1.17.0", "types-protobuf>=6.30.2.20250822"]
testing = ["pytest>=8.3.4"]
examples = [
    "langchain-core",
    "langchain-openai",
    "typing-extensions",
    "uvicorn>=0.34.2",
]

# Default groups to install when running `uv sync`
[tool.uv]
default-groups = ["linting", "testing", "examples"]

[tool.ruff.format]
exclude = ["*_pb2.py", "*_pb2.pyi", "*_pb2_slimrpc.py"]

[tool.ruff.lint]
extend-select = ["I", "ASYNC", "SIM", "B", "ANN"]
exclude = ["*_pb2.py", "*_pb2.pyi", "*_pb2_slimrpc.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["slima2a"]
exclude = ["examples"]

[tool.hatch.build.targets.sdist]
exclude = ["examples"]

[[tool.mypy.overrides]]
module = ["google.rpc.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["slim_bindings.*"]
follow_untyped_imports = true
#ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["slima2a.types.v0.a2a_pb2_slimrpc"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["slima2a.types.v1.a2a_pb2_slimrpc"]
ignore_errors = true
